+++ script/module/html.js
@@ -176,8 +176,14 @@ window.theme.openNewWindow = function (
             console.warn(err);
             // 新建标签页(Web 环境)
             // window.open(url.href, "_blank");
-            window.open(url.href, "popup");
-            return null;
+            // REF [Window.open() - Web APIs | MDN](https://developer.mozilla.org/en-US/docs/Web/API/Window/open)
+            // REF [Window open() 方法 | 菜鸟教程](https://www.runoob.com/jsref/met-win-open.html)
+            newWin = window.open(url.href, url.href, `
+                popup = true,
+                width = ${windowParams.width},
+                height = ${windowParams.height},
+            `);
+            return newWin;
         }
     }
     catch (err) {